Give Sippy chat read-only access to the postgresql database
This opens up the possibility to answer varied questions about jobs,
tests, etc without writing custom tools to do it.  With some coaching
about the database schema, and some notes about its warts, it seems
pretty good (with some thrashing) to answer questions like:

  - How are metal ipv6 jobs doing in 4.17?
  - How did single node jobs do in the last 24 hours?
  - Can you see if GCP jobs are doing worse in 4.21 vs 4.20 in the last week?
  - Can you analyze the test failures for "test" in single node jobs?
  - What were the last 5 outputs for failures of "test"?

This works very well when supplying a small amount of context on job and
test analysis pages.

A number of controls are in place:
  - The LLM is instructed to only answer questions about CI
  - Read-only is enforced with multiple fail-safe:
      - by the user permissions we'll use (sippyro)
      - the client's session (read_only=True)
      - filtering on verbs (no INSERT, DELETE, etc).
  - A blocklist of tables is provided, e.g. no `pg_*` tables are
    accessible