crcacheloader: cache test_details even with errors
When anything goes wrong in one test_details report, an error is
returned and the report is not added to the list. Before this change,
when that error was returned in the cache-primers multi-test report,
this was immediately returned and thus caching was skipped for the
entire view, even though most results were successful.

With this change, errors are collected and any actual results are cached
before returning them. We assume that any results returned are
successful - should not return reports that are tainted by error.