CR main: still show data with 1 row
For some reason, at the top-level/main report, if the result only comes
back with one row, this logic erased the columns. This is what I saw the
first time I restricted Capabilities and it was very puzzling to see
regressions but no data. I cannot imagine why we would not want to show
the data in that case, so I changed the logic to show any data there is.