Try to convince Gemini to stop modifying URL's
Gemini does not like Component Readiness URL's and gets confused by
dbGroupBy and the list of included variants, to the point it often
modifies the URL by putting dbGroupBy=Network:ovn instead of what was
originally in the URL (dbGroupBy=Network).

Example:
  https://sippy-auth.dptools.openshift.org/sippy-ng/chat/b0265da2-89db-4baf-9569-7ac74954c6dd

This tries to convince Gemini to stop modifying things to tools that ask
for verbatim parameters.