jobartifacts: pass request ctx to GCS lookup
This was a suggestion from CodeRabbit that seemed to make sense. If the
request context gets canceled, then the results are going to be
abandoned anyway, so it makes sense to let that cancellation pass to the
GCS reader and give it a chance to abort.

It seems right... It should be a minor optimization... But check if
anything breaks.